feat(git): Refactored GitDetectable to GitCliDetectable, stopped parsing in the GitCliDetectable.
